+++ src/Application.php
@@ -12,7 +12,10 @@ use Innmind\CLI\{
     Environment as CliEnv,
     Command,
 };
-use Innmind\DI\Container;
+use Innmind\DI\{
+    Container,
+    Service,
+};
 use Innmind\Http\{
     ServerRequest,
     Response,
@@ -107,12 +110,12 @@ final class Application
     /**
      * @psalm-mutation-free
      *
-     * @param non-empty-string $name
+     * @param non-empty-string|Service $name
      * @param callable(Container, OperatingSystem, Environment): object $definition
      *
      * @return self<I, O>
      */
-    public function service(string $name, callable $definition): self
+    public function service(string|Service $name, callable $definition): self
     {
         return new self($this->app->service($name, $definition));
     }