+++ src/Application/Async/Http.php
@@ -19,6 +19,7 @@ use Innmind\Async\HttpServer\Command\Serve;
 use Innmind\DI\{
     Container,
     Builder,
+    Service,
 };
 use Innmind\Http\{
     ServerRequest,
@@ -137,7 +138,7 @@ final class Http implements Implementation
     /**
      * @psalm-mutation-free
      */
-    public function service(string $name, callable $definition): self
+    public function service(string|Service $name, callable $definition): self
     {
         return new self(
             $this->os,