+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Support for using enums as a service name
+
 ## 2.1.0 - 2024-03-10
 
 ### Added